Chore: rename wrapper to net
