+++ common/net/io.go
@@ -1,4 +1,4 @@
-package wrapper
+package net
 
 import "io"
