+++ tunnel/connection.go
@@ -10,8 +10,8 @@ import (
 	"time"
 
 	"github.com/Dreamacro/clash/adapters/inbound"
+	N "github.com/Dreamacro/clash/common/net"
 	"github.com/Dreamacro/clash/common/pool"
-	"github.com/Dreamacro/clash/common/wrapper"
 	"github.com/Dreamacro/clash/component/resolver"
 	C "github.com/Dreamacro/clash/constant"
 	"github.com/Dreamacro/clash/context"
@@ -142,14 +142,14 @@ func relay(leftConn, rightConn net.Conn) {
 
 	go func() {
 		buf := pool.Get(pool.RelayBufferSize)
-		_, err := io.CopyBuffer(wrapper.WriteOnlyWriter{Writer: leftConn}, wrapper.ReadOnlyReader{Reader: rightConn}, buf)
+		_, err := io.CopyBuffer(N.WriteOnlyWriter{Writer: leftConn}, N.ReadOnlyReader{Reader: rightConn}, buf)
 		pool.Put(buf)
 		leftConn.SetReadDeadline(time.Now())
 		ch <- err
 	}()
 
 	buf := pool.Get(pool.RelayBufferSize)
-	io.CopyBuffer(wrapper.WriteOnlyWriter{Writer: rightConn}, wrapper.ReadOnlyReader{Reader: leftConn}, buf)
+	io.CopyBuffer(N.WriteOnlyWriter{Writer: rightConn}, N.ReadOnlyReader{Reader: leftConn}, buf)
 	pool.Put(buf)
 	rightConn.SetReadDeadline(time.Now())
 	<-ch
